move tabs into a seperate component
