+++ src/components/EditorView.js
@@ -2,14 +2,14 @@ import React, { Fragment } from "react";
 
 import Editor from "./Editor";
 import DetectPaste from "./DetectPaste";
-import Transformation from "./Transformation";
+import Tabs from "./CollapsibleTabs/Tabs";
 
 const EditorView = () => {
   return (
     <Fragment>
       <Editor />
+      <Tabs />
       <DetectPaste />
-      <Transformation />
     </Fragment>
   );
 }
