+++ src/components/Transformation/index.js
@@ -1,30 +1,18 @@
 import React from "react";
 
-import CollapsibleTabs from "../CollapsibleTabs";
 import getEditor from "../../utils/editor-actions";
 import Editor from "../Editor";
 
 function Transformation() {
   const transformEditor = getEditor();
-  const tabs = [
-    { key: "transformation", header: <span>Transform</span>}
-  ];
-  const createComponent = (selected) => {
-    switch(selected) {
-      case "transformation":
-        return (
-          <Editor modelType="js" editorConfig={transformEditor} />
-        );
-      default:
-        return null;
-    }
-  }
   return (
-    <CollapsibleTabs
-      tabs={tabs}
-    >
-      {createComponent}
-    </CollapsibleTabs>
+    <div css={`height: 100%;`}>
+      <header>
+        <h2>Transformation</h2>
+        <button>Run</button>
+      </header>
+      <Editor modelType="js" editorConfig={transformEditor} />
+    </div>
   );
 }
